+++ statistics.cpp
@@ -7,7 +7,7 @@ float mean_velocity(std::vector<Boid> const& stormo)
   for (unsigned long int i{0}; i != stormo.size(); ++i) {
     sum += distanza(stormo[i].get_vel(), {0, 0});
   }
-  return sum / (stormo.size());
+  return sum / static_cast<float>((stormo.size()));
 }
 
 float mean_position(std::vector<Boid> const& stormo)
@@ -16,27 +16,27 @@ float mean_position(std::vector<Boid> const& stormo)
   for (unsigned long int i{0}; i != stormo.size(); ++i) {
     sum += distanza(stormo[i].get_pos(), {0, 0});
   }
-  return sum / (stormo.size());
+  return sum / static_cast<float>((stormo.size()));
 }
 
 float standdev_velocity(std::vector<Boid> const& stormo)
 {
   float sum{0};
   for (unsigned long int i{0}; i != stormo.size(); ++i) {
-    sum +=
-        powf((distanza(stormo[i].get_vel(), {0, 0}) - mean_velocity(stormo)), 2);
+    sum += powf((distanza(stormo[i].get_vel(), {0, 0}) - mean_velocity(stormo)),
+                2);
   }
-  return std::sqrtf(sum / (stormo.size() - 1));
+  return std::sqrt(sum / static_cast<float>((stormo.size() - 1)));
 }
 
 float standdev_position(std::vector<Boid> const& stormo)
 {
   float sum{0};
   for (unsigned long int i{0}; i != stormo.size(); ++i) {
-    sum +=
-        powf((distanza(stormo[i].get_pos(), {0, 0}) - mean_position(stormo)), 2);
+    sum += powf((distanza(stormo[i].get_pos(), {0, 0}) - mean_position(stormo)),
+                2);
   }
-  return std::sqrtf(sum / (stormo.size() - 1));
+  return std::sqrt(sum / static_cast<float>((stormo.size() - 1)));
 }
 
 float mean_distance(std::vector<Boid> const& stormo)
@@ -58,11 +58,11 @@ float standdev_distance(std::vector<Boid> const& stormo)
   for (unsigned long int i{0}; i != stormo.size(); ++i) {
     for (unsigned long int j{0}; j != i; ++j) {
       sum += powf((distanza(stormo[i].get_pos(), stormo[j].get_pos())
-                  - mean_distance(stormo)),
-                 2);
+                   - mean_distance(stormo)),
+                  2);
     }
   }
-  return std::sqrtf(sum
+  return std::sqrt(sum
                    / ((static_cast<float>(stormo.size())
                        * (static_cast<float>(stormo.size()) - 1))
                           / 2.0f
